+++ lib/src/services/api.service.dart
@@ -133,9 +133,7 @@ class ApiService {
       return;
     }
     reconnectionTimer?.cancel();
-    Log.info('Starting reconnection timer with $_reconnectionDelay s delay');
     reconnectionTimer = Timer(Duration(seconds: _reconnectionDelay), () async {
-      Log.info('Reconnection timer triggered');
       reconnectionTimer = null;
       // only try to reconnect in case the app is in the foreground
       if (!globalIsAppInBackground) {